# RustGameAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RustGameAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Friend/AddFriend": {
      "post": {
        "tags": [
          "Friend"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FriendDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Friend/SendFriendRequest": {
      "post": {
        "tags": [
          "Friend"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "friendUsername",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Friend/{userId}": {
      "get": {
        "tags": [
          "Friend"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Friend/GetAllFriends/{userId}": {
      "get": {
        "tags": [
          "Friend"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/AddUser": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetAllUsers": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/LookForUser": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetIDfromUsername/{Username}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetScoresDescending": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/UpdateHighScore": {
      "put": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UserID",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "HighScore",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetHighScore/{UserID}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserID",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/ChangeUsername": {
      "put": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UserID",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "NewUsername",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/ChangePassword": {
      "put": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UserID",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "NewPassword",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "DateOnly": {
      "type": "object",
      "properties": {
        "Year": {
          "format": "int32",
          "type": "integer"
        },
        "Month": {
          "format": "int32",
          "type": "integer"
        },
        "Day": {
          "format": "int32",
          "type": "integer"
        },
        "DayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "DayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "DayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
    "DayOfWeek": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "FriendDTO": {
      "type": "object",
      "properties": {
        "UserID": {
          "format": "int32",
          "type": "integer"
        },
        "FriendUserID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "Date": {
          "$ref": "#/definitions/DateOnly"
        },
        "TemperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "TemperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "Summary": {
          "type": "string"
        }
      }
    }
  }
}